+++ config/queue.php
@@ -99,6 +99,27 @@ return [
             'worker' => env('RABBITMQ_WORKER', 'default'),
         ],
 
+        'rabbitmq-external-remains' => [
+            'driver' => 'rabbitmq',
+            'queue' => env('RABBITMQ_EXTERNAL_REMAINS_QUEUE', 'external.remains_for_website'),
+            'connection' => PhpAmqpLib\Connection\AMQPLazyConnection::class,
+            'hosts' => [
+                [
+                    'host' => env('RABBITMQ_HOST', 'rabbitmq'),
+                    'port' => env('RABBITMQ_PORT', 5672),
+                    'user' => env('RABBITMQ_USER', 'guest'),
+                    'password' => env('RABBITMQ_PASSWORD', 'guest'),
+                    'vhost' => env('RABBITMQ_VHOST', '/'),
+                ],
+            ],
+            'options' => [
+                'queue' => [
+                    'job' => App\Queue\Jobs\ExternalRemainsJob::class,
+                ],
+            ],
+            'worker' => env('RABBITMQ_WORKER', 'default'),
+        ],
+
         'beanstalkd' => [
             'driver' => 'beanstalkd',
             'host' => env('BEANSTALKD_QUEUE_HOST', 'localhost'),